Write the document level to aux for reference
We will want to reference the marking level of the overall document when
formatting a cover page.  This means we need the final level at the
start of the second pass.  This means we need to write out the final
marking at the end of the document to the auxiliary file.