Remove restriction on nested markings
It turns out we can't enforce the no nested marking rule in a general
document.  I figured that out trying to use this on a real document.
Instead, we need to store the level at the start of the macro and reset
it after we're done.  This requires some trickery with grouping the
macros and defining a nice way to get `\aftergroup` to reset the current
level counter correctly.