Conditionally set the header and footer
We have two conditions in which we want to user markings.  Articles and
presentations.  For articles, we want to keep doing what we're doing and
use fancy headers.  For beamer slides, we want to use the headline and
footline.  However, we expect the user to want to use a fancier
formatting, so we provide a very simple marking.