Reformat additions to match the main document
Basically, put the description inside a `\macro` environment to get the
margin call out, reformat to 72 columns, and relabel the section title
to match what it actually is.  While at it, I changed the name from
`manheading` to `heading` because that's not in use and a clearer
variable name, dropped the need for iftex, and judiciously added '%' to
ensure no additional line breaks will creep in.