Add a build.lua for use with l3build
The l3build tool provides a much more streamlined method to generate the
TDS than the dtx Tool for SCons I had been working on.  It's much more
concise, builds the TDS and CTAN format, and will make it easy to
upload.  It also has a means for establishing tests which will hopefully
help with regression checking.