Add tests of individual marking macros
These are the tests to check that running text gets marked at the start
of each paragraph with a marking both in a macro and in an environment.
Each test checks a single level.

If run with an individual engine (l3build check -e {pdftex,xetex,luatex})
they all pass.  However, running without an engine indicates xetex and
luatex fail.  This is because some details vary between engines (thus
three versions of the logs).  I suspect the bulk run can be corrected
with an input in the build.lua, but I don't know what it is yet.  For
now, just run engines individually.