Edit the description to better explain the code
The description of how and why we were using PushPostHook (including
where it comes from) and the adjusting of the list paragraph indentation
wasn't very clear.  Instead of having to try to figure out what it's
doing again, the better answer it to add some more real text explaining
what is going on.