Use the label width to set the list markings
Instead of explicitly setting the label width with `enumitem`, we can
just use the `\labelwidth` (native to LaTeX) as the padding to `\kern`.
This removes a dependency (always a good thing).  A side effect is the
paragraphs in a list wind up being better aligned than the original
method.  This doesn't fix the problem with the description environment,
but it pushes us in the proper direction.