Separate the mark positioning code to a macro
Instead of having to duplicate the details of the horizontal positioning
code, we just put it in a separate macro.  I know this is really short
and seems like it's not much, but I'm starting to suspect that the
details of the positioning will become more complicated as we try to
figure out where the margin actually is and place the marking just
outside the margin.