Define a fancyhdr page style
Instead of completely overriding the `fancy` page style, we should
listen to the documentation and define a customer page style.  Further,
we should update to the newer syntax to be forward portable.  This also
removes the page number from the style because that should be handled by
some other component of the layout.