Use the recommended method in the everyhook document
The everyhook package recommends placing the item we want into a box and
toggling around the box order to keep it in the right location.  This
also means we need to adjust the spacing around the lists.  But, it does
put the main markings in the margin like I originally wanted.